Do not trim nicknames at ~
First, it's 2013 and I'd be really surprised if anybody still cared
about ISO-646 support these days.

Second, ISO-646-FI, which the "scandinavian origins" probably refers to,
did NOT have any letters in place of ~ or ^; it had the same ~ ^ and so
did all other Scandinavian ISO-646 charsets. The only exception here is
ISO-646-SE-C, which had Ü ü in those places, but it also had É é in
place of @ `, so by this logic the code should disallow ` as well.

Fuck that. If this patch hurts ISO-646 users, that's a feature.