Granular Role-Based Access Controls. This comprehensively shifts access delegation away from developers entirely into the administrator's configuration. The technical goal is to render the setting of umode +o ipso facto negligible.
The flexibility of this system is based on the ability to compose
a hierarchy similar to privset but with every permitted command path,
chmode, umode, snomask, and exemption enumerated from within the
configuration. When an operator is not permitted, the behavior is to
transparently fall back to that of a normal user.

This supercedes the privset configuration with a more flexible role
configuration. The role {} newconf block now takes over for deprecated
privset {}. This replaces all relevant calls to IsOper() and all
calls to privset's IsOper*() with the appropriate role call instead.