Granular RBAC: Preliminary Exemption Mode Masking. Transforms the coagulation of various exemption flags and macros riding client->flags2 into a characterized mode-mask. The flags2 overflow is removed, and an additional non-exempt related flags2 was moved to flags; 16bit architectures may no longer be supported.
This further completes the ability to finely configure access rather
than coarse calls to IsOper() or a few arbitrary exemption macros. For
example, an operator with a limited role can't flood NICK changes just
because of +o. Instead they would require flood exemption (MODE +e +f),
which is still somewhat coarse for me but this diff is big enough for
now.

The UMODE_EXEMPT (+e) is introduced and behaves like UMODE_SERVNOTICE
(+s), However, there is no server protocol support yet for parameters
to any usermodes so exemptions are not honored by remote servers with
this commit.