Granular RBAC: Configuration. The mode specifications for roles increase in their dimensionality, where only the first in the following list was available formerly:
* AVAIL   - The role has the mode available to them.
* DEFAULT - The mode is switched on during oper_up.
* LOCKED  - The mode cannot be switched off.

This allows the removal of several redundancies in the configuration
and elsewhere for modes an operator gets by default, etc. Now this
is all done through one place: the role.

Additionally some more complex modifications to s_user provide
completed functionality left out of previous commits to simplify
their diffs for audit.

As of this commit the entire mask under umode +e is honored in full
by remote servers, and +e is set in full when remote clients go +o.