Modules in Hausdorff space. Facilitated by MAPIv3, we introduce modules that moonlight as actual subprocesses. This is achieved by inserting an elf header entry point into the module object and executing it. The module will then operate disjoint address spaces with asynchronous message passing between.
The advantage is that all code, callout, callback, child handlers,
whatever, can all exist and be controlled near each other. When the
module is loaded, everything begins; when it is unloaded, everything
finishes.

Subprocesses already have the advantage of being able to safely crash
and restart, there is simply no uniform interface to manage and inter-
act with them, to manage state between callbacks, message identities,
crash recovery, etc. This provides that interface.