GNUTLS: Store error codes properly
This is similar to commit db12df5c16758487e0c88d43cd77926b784fe400 for
the MbedTLS backend.

The difference is, GNUTLS will not accept positive values as input to
gnutls_strerror(), so we invert the sign bit after retrieving the value
too, not just when storing it.

Also add a forgotten ssl_errno assignment to rb_ssl_connect_common().